requirements.txt
asgiref==3.5.2
Django==3.2.15
djangorestframework==3.13.1
importlib-metadata==4.12.0
Markdown==3.4.1
Pillow==9.2.0
psycopg2==2.9.3
pytz==2022.2.1
sqlparse==0.4.2
typing_extensions==4.3.0
zipp==3.8.1
